Rename shared and managed to inherited and shared.
